requirements.txt
cryptography==41.0.4
pyjwt==2.8.0
requests==2.33.0
